Move step ca sign code to sign.go
